Add ready regex to the popup list & cleanup html
The list for whitelisting/greylisting will now include a third option,
which is a regex pattern that matches the domain. This is the whole
purpose of this fork; I use regex for my expressions, and this makes it
so much easier to maintain. It also puts the domain in parentheses, in
case multiple domains are wanted in the same expression.